P14C-622: Refactored user info. component
